step3 : racing car - 리펙토링
