--- src/main/kotlin/racing/constant/RacingGameConstant.kt
@@ -1,7 +1,0 @@
-package racing.constant
-
-object RacingGameConstant {
-    const val JOIN_CAR_COUNT_QUESTION = "자동차 대수는 몇 대인가요?"
-    const val TRY_GAME_COUNT_QUESTION = "시도할 횟수는 몇 회인가요?"
-    const val RESULT_SENTENCE = "실행 결과"
-}
