+++ src/main/kotlin/racing/model/Racing.kt
@@ -1,0 +1,5 @@
+package racing.model
+
+class Racing(carCount: Int) {
+    var cars: List<Car> = List(carCount) { Car() }
+}
